fixed int/double precision loss

+++ simulator/RobotAPI.java
@@ -10,9 +10,9 @@
 public interface RobotAPI {
 
   // moves the robot in a straigth line for a distance expressed in meters
-  public void move( float distance );
+  public void move( double distance );
   
   // turns the robot on its spot by an angle expressed in degrees
-  public void turn( int angle );
+  public void turn( double angle );
   
 }
